Renamed flash_config to bl_flash_config
